fix 826: Spatial Filtering grid definition
The grid was constructed using np.meshgrid over the range of the bounding box,
with a computed stride. If the range and number of divisions were both integer,
as was the case in the unittest, this generated floor division incorrectly. When
we did initial testing in python3, we thought the unittest was correct. In fact,
the unittest was incorrect, and we needed to enforce float division always.