Handle slopes of key cols
When a key col is found, it has to join *new* peaks to the graph.

For example, say we're stepping down the ranks. We start with peaks
A, B, C.

We find a key col between A and B. Next, we find a key col between
A and C. Then, when we find a col between B and C, they're
*already* joined through A. The predecessors go

A
A <- B
(A <- B) <- C
So, B <- C is a connection *within* the A <- B <- C graph.