Correct documentation
Setting transformation to "U" causes an "unsupported weights transformation" exception because "U" is not an accepted parameter of the set_transform method of the libpysal.weights.W class. The correct argument is "O", which is described in the libpysal.weights.W documentation as "Restore original transformation (from instantiation)". I've updated the documentation to reflect that.